fix(telemetry): Tested and fixed sending telemetry issue in the logic

+++ src/cores/core_1/feature/log/telemetry/json_builder/json_builder.c
@@ -122,12 +122,12 @@ const char* json_get(const Json_h* const restrict self)
   return p_self->p_str_json;
 }
 
-uint16_t json_len(const Json_h* const restrict self)
+uint32_t json_len(const Json_h* const restrict self)
 {
   const union Json_h_t_conv_const conv = {self};
   const struct Json_t* const p_self = conv.clear;
 
-  return (uint16_t) strlen(p_self->p_str_json);
+  return (uint32_t) strlen(p_self->p_str_json);
 }
 
 int8_t json_destroy(Json_h* self)
